# WIP: directed_hausdorff speedup attempt