This WIP PR attempts to improve the performance of `directed_hausdorff` by abstracting the heavy-lifting to an external C module. It was admittedly motivated by an attempt to use some newly-acquired knolwedge of the C language & see what (if any) performance benefits might be achieved. So far, only very modest improvements have been achieved & the increase in complexity / maintenance burden is likely not justified at this stage.

Perhaps some of the C gurus will have some additional suggestions? I'll put a few notes below. We can always just close / archive this in case there's interest in the C version for whatever reason later.

C code notes:
- uses the controversial `goto` statement & there wasn't really much performance benefit for adding this (to circumvent a boolean variable + check for nested loop continuation) -- could refactor to a function that uses a `return` perhaps or just restore the original boolean variable and `break` statement
- there's likely no compelling reason to use so much pointer arithmetic -- this was again admittedly mostly for the learning experience, and there may even be drawbacks to that (vs. conventional array indexing) as these details may best be left to the compiler
- line profiling the Cython `pyx` + associated C header file seems to be a bit painful -- would be really nice if this were much simpler to drill right down to the C code lines in parallel with Cython lines; I tried various permutations of `%%cython` magic in Jupyter notebooks and `lprof` and so on, but each approach seems to have drawbacks or require quite a bit of manual intervention
- overall, Cython does seem to do a really good job & the ROI for time writing better raw C code so far seems questionable

Sample benchmark comparison with latest commit hash vs. master branch point:
`asv continuous --bench Hausdorff 355a8e08fb 38eb3e47d`

```
· Creating environments
· Discovering benchmarks
·· Uninstalling from conda-py3.6-Cython0.26-Tempita-numpy-pytest-six
·· Installing into conda-py3.6-Cython0.26-Tempita-numpy-pytest-six
· Running 2 total benchmarks (2 commits * 1 environments * 1 benchmarks)
[  0.00%] · For scipy commit hash 38eb3e47:
[  0.00%] ·· Building for conda-py3.6-Cython0.26-Tempita-numpy-pytest-six
[  0.00%] ·· Benchmarking conda-py3.6-Cython0.26-Tempita-numpy-pytest-six
[ 50.00%] ··· Running spatial.Hausdorff.time_directed_hausdorff               ok
[ 50.00%] ···· 
               ============ ============
                num_points
               ------------ ------------
                    10       26.9±0.1μs
                   100       96.2±0.2μs
                   1000       1.11±0ms
                  10000      24.6±0.2ms
                  100000      1.43±0s
               ============ ============

[ 50.00%] · For scipy commit hash 355a8e08:
[ 50.00%] ·· Building for conda-py3.6-Cython0.26-Tempita-numpy-pytest-six
[ 50.00%] ·· Benchmarking conda-py3.6-Cython0.26-Tempita-numpy-pytest-six
[100.00%] ··· Running spatial.Hausdorff.time_directed_hausdorff               ok
[100.00%] ···· 
               ============ ============
                num_points
               ------------ ------------
                    10       28.1±0.1μs
                   100        104±1μs
                   1000       1.23±0ms
                  10000      26.6±0.4ms
                  100000     1.53±0.01s
               ============ ============
       before           after         ratio
     [355a8e08]       [38eb3e47]
-        1.23±0ms         1.11±0ms     0.90  spatial.Hausdorff.time_directed_hausdorff(1000)

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.

```
ping @kain88-de @orbeckst for possible interest