@tylerjereddy 2017-11-21
Ok, closing because insufficient improvements in performance relative to the amount of extra C code added, etc. However, I may try to bump the Cython performance slightly, as discussed above, in a separate smaller PR.

This can still serve as a useful reference C implementation of much of the hausdorff code.