@tylerjereddy 2017-11-21
No evidence of substantial speedup when making the change discussed above in Cython.